Add files via upload
Sara and Josh are participating in a treasure hunt where the hints are hidden in the meanings of different words. You have created an online Dictionary App to help Josh and Sara find the meanings of the words.

Now you have to create a dictionary app which will work without an internet connection.